Set kTruncatingToInt32 flag if one of the SIMD operand requires Integer32
Before this, we deoptimize for SIMD.int32x4.splat(0x80000000) and
SIMD.int32x4(0x80000000, 0xFFFFFFFF, 0xFFFFFFFF, 0xFFFFFFFF) as we did
not truncate the "double/unsigned int32" to signed int32. It seems OK to
set kTruncatingToInt32 flag in a general way so it works for int32x4
shift and shuffle operation also.