Crankshaft disable option for CPU profiler
This patch allows to increase the accuracy of the source code annotations by
profiling metrics. To achieve this we turns off the crankshaft compiler
and deoptimize already compiled code.

Patch will allow to investigate the performance of algorithms on the
source level, not in the function modes. In spite the final code during
profiling will be different than working in regular mode, this feature
still makes sense especially because similar example of turning off
crankshaft compiler – running of the code under different JavaScript
engine. It will not allow to make m-arch analysis for JavaScript code
working under v8 and make m-arch optimizations but this feature will be
useful for investigations of performance on algorithm level.

To give the ability of the original code analyzing, we are enabling above
mode through the new option for the cpu profiling start. If the option is
set to false, the behavior will be absolutely the same as it is now –
heavy optimized code with problems in debug info.

Also, this patch disables crankshaft during XDK Heap Profiler collection.