Adjust coding style in files touched by e236921.
The presubmit hooks fail even though the original upstream files are not
compliant. Run `git-cl format v8` to make the coding style checks pass.
This commit was not squashed into e236921 (Revert "XDK Heap Profiler
collector and XDK CPUProfiler") to make it easier for future rebases to
skip this change if necessary.